% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedPValues.R
\name{CombinedPValueMethod}
\alias{CombinedPValueMethod}
\title{CombinedPValueMethod}
\usage{
CombinedPValueMethod(x, name)
}
\arguments{
\item{x}{#' test statistic of method used (i.e., Tippett, Stouffer, etc.)}

\item{name}{# name of method using}
}
\value{
Combined P-value
}
\description{
#' Input is the test statistic of the previous method selected and it returns the combined p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- TippettMethod(Output)
Combined <- CombinedPValueMethod(Final,"Tippett")
}
