% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgingtonMethod.r
\name{EdMethod}
\alias{EdMethod}
\title{Edgington Method}
\usage{
EdMethod(x)
}
\arguments{
\item{x}{#' SumOfPs}
}
\value{
Combined P-value
}
\description{
#' Combination p-value method that uses Edgington statistic
Summation i=1 to n pi where p equals p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- EdMethod(Output)
}
