% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishersMethod.r
\name{FishersMethod}
\alias{FishersMethod}
\title{FishersMethod}
\usage{
FishersMethod(x)
}
\arguments{
\item{x}{#' SumOfPs}
}
\value{
Combined P-value
}
\description{
#' Combination p-value method that uses Fishers statistic
Summation i=1 to n log of pi where p equals p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- FishersMethod(Output)
}
