% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeorgeMethod.r
\name{GeorgeMethod}
\alias{GeorgeMethod}
\title{PearsonsMethod}
\usage{
GeorgeMethod(x)
}
\arguments{
\item{x}{#' SumOfPs}
}
\value{
Combined P-value
}
\description{
#' Combination p-value method that uses George statistic
Summation i=1 to n log(pi/(1-pi)) where p equals p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- GeorgeMethod(Output)
}
