% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfinitePs.r
\name{InfinitePs}
\alias{InfinitePs}
\title{InfinitePs}
\usage{
InfinitePs(x, ...)
}
\arguments{
\item{x}{#' Input n p-values  n = 2,3,...,k}

\item{...}{#list of p values}
}
\value{
List of p-values
}
\description{
Converts a list of p-values into a list, n= 2,3,...,k
}
\examples{
Output <- InfinitePs(0.1,0.3,.7)
}
