% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PickMethod.r
\name{PickMethod}
\alias{PickMethod}
\title{PickMethod}
\usage{
PickMethod(x)
}
\arguments{
\item{method}{A data set object}
}
\value{
Allows for the proper function to be selected
}
\description{
function that allows the user to define which method of combination p-value 
they want use
}
\examples{
my_object <- PickMethod("Perason")
}
