% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoufferMethod.r
\name{StoufferMethod}
\alias{StoufferMethod}
\title{StoufferMethod}
\usage{
StoufferMethod(x)
}
\arguments{
\item{x}{#' SumOfPs}
}
\value{
Combined P-value
}
\description{
#' Combination p-value method that uses Stouffer statistic
Summation i=1 to n inverse CDF of N(0,1)(pi) where p equals p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- StoufferMethod(Output)
}
