% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TippettMethod.r
\name{TippettMethod}
\alias{TippettMethod}
\title{TippettMethod}
\usage{
TippettMethod(x)
}
\arguments{
\item{x}{#' SumOfPs}
}
\value{
Combined P-value
}
\description{
#' Combination p-value method that uses Tippett statistic
min(p1,...,pn), n= 2,3,...,k where p equals p-value
}
\examples{
Output <- SumOfPs(0.1,0.3,.7)
Final <- TippettMethod(Output)
}
